# release-24.1: sqlstats: increase in-memory fingerprint limit from 5000 to 7500