… docs/generated/settings/settings.html