Backport 1/1 commits from #125343.

/cc @cockroachdb/release

---

Previously we set the in-memory fingerprints for sql stats to be capped at 5000 each for statemnts and transactions. We've found that 7500 is a more suitable number that will prevent dropped stats for most workloads we have this data for.

Before:
BenchmarkSqlStatsMaxFlushTime/single-application/writes=insert/10000-fingerprints-10   1   13005720666 ns/op
BenchmarkSqlStatsMaxFlushTime/single-application/writes=update/10000-fingerprints-10   1   15769446916 ns/op

After:
BenchmarkSqlStatsMaxFlushTime/single-application/writes=insert/15000-fingerprints-10   1   19813686375 ns/op
BenchmarkSqlStatsMaxFlushTime/single-application/writes=update/15000-fingerprints-10   1   23514272626 ns/op

Release note: None

Epic: CRDB-37544